+++ packages/site/src/components/Navigation/Navigation.tsx
@@ -4,13 +4,13 @@ import { TableOfContents } from './TableOfContents.js';
 export function Navigation({
   children,
   projectSlug,
-  tocRef,
+  sidebarRef,
   hide_toc,
   footer,
 }: {
   children?: React.ReactNode;
   projectSlug?: string;
-  tocRef?: React.RefObject<HTMLDivElement>;
+  sidebarRef?: React.RefObject<HTMLDivElement>;
   hide_toc?: boolean;
   footer?: React.ReactNode;
 }) {
@@ -30,7 +30,7 @@ export function Navigation({
           onClick={() => setOpen(false)}
         ></div>
       )}
-      <TableOfContents tocRef={tocRef} projectSlug={projectSlug} footer={footer} />
+      <TableOfContents sidebarRef={sidebarRef} projectSlug={projectSlug} footer={footer} />
       {children}
     </>
   );