+++ packages/frontmatter/src/FrontmatterBlock.tsx
@@ -74,9 +74,16 @@ export function DateString({
   spacer?: boolean;
 }) {
   if (!date) return null;
-  const d = new Date(date); // This is in the users timezone
-  const utcDate = new Date(d.getUTCFullYear(), d.getUTCMonth(), d.getUTCDate());
-  const dateString = utcDate.toLocaleDateString('en-US', format);
+  // Parse the date
+  // As this is a YYYY-MM-DD form, the parser interprets this as a UTC date
+  // (see https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Date#the_epoch_timestamps_and_invalid_date)
+  const utcDate = new Date(date);
+
+  // Now cast our UTC-date into the local timezone
+  const localDate = new Date(utcDate.getUTCFullYear(), utcDate.getUTCMonth(), utcDate.getUTCDate());
+
+  // Then format as human-readable in the local timezone.
+  const dateString = localDate.toLocaleDateString('en-US', format);
   return (
     <time dateTime={date} className={classNames({ 'text-spacer': spacer })}>
       {dateString}
@@ -228,7 +235,7 @@ export function FrontmatterBlock({
       className={classNames(className)}
     >
       {showHeaderBlock && (
-        <div className="flex items-center h-6 mt-3 mb-5 text-sm font-light">
+        <div className="flex items-center h-6 mb-5 text-sm font-light">
           {subject && (
             <div
               className={classNames('flex-none pr-2 smallcaps', {