+++ themes/book/remix.config.dev.js
@@ -61,9 +61,11 @@ module.exports = {
   ],
   watchPaths: ['../../packages/**/*'],
   future: {
+    v2_dev: true,
     v2_routeConvention: true,
     v2_normalizeFormMethod: true,
     v2_headers: true,
     v2_meta: true,
+    v2_errorBoundary:true,
   },
 };